+++ app/_layout.tsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react';
 import { SplashScreen, Stack } from 'expo-router';
 import { useFonts } from 'expo-font';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { TokenProvider } from '../contexts/TokenContext';
 import Toast, { BaseToast, ErrorToast } from 'react-native-toast-message';
 import { NativeWindStyleSheet } from 'nativewind';
 import { tokens } from '../constants';
@@ -50,29 +51,31 @@ const RootLayout = () => {
   
   return (
     <QueryClientProvider client={queryClient}>
-      <Stack screenOptions={{
-        headerTitleAlign: 'center',
-      }}>
-        <Stack.Screen name="index" options={{ headerShown: false }}/>
-        <Stack.Screen name="(auth)/sign-in" options={{
-          animation: 'slide_from_bottom',
-          headerShown: false,
-        }} />
-        <Stack.Screen name="(auth)/sign-up" options={{
-          animation: 'slide_from_right',
-          title: '회원가입',
-        }} />
-        <Stack.Screen name="(tabs)" />
-        <Stack.Screen name="screens/CreateGroup" options={{
-          title: '모임 등록',
-          animation: 'slide_from_bottom',
-        }} />
-        <Stack.Screen name="screens/CreateDiving" options={{
-          title: '다이빙 등록',
-          animation: 'slide_from_bottom',
-        }}/>
-      </Stack>
-      <Toast config={toastConfig}/>
+      <TokenProvider>
+        <Stack screenOptions={{
+          headerTitleAlign: 'center',
+        }}>
+          <Stack.Screen name="index" options={{ headerShown: false }}/>
+          <Stack.Screen name="(auth)/sign-in" options={{
+            animation: 'slide_from_bottom',
+            headerShown: false,
+          }} />
+          <Stack.Screen name="(auth)/sign-up" options={{
+            animation: 'slide_from_right',
+            title: '회원가입',
+          }} />
+          <Stack.Screen name="(tabs)" />
+          <Stack.Screen name="screens/CreateGroup" options={{
+            title: '모임 등록',
+            animation: 'slide_from_bottom',
+          }} />
+          <Stack.Screen name="screens/CreateDiving" options={{
+            title: '다이빙 등록',
+            animation: 'slide_from_bottom',
+          }}/>
+        </Stack>
+        <Toast config={toastConfig}/>
+      </TokenProvider>
     </QueryClientProvider>
   );
 };
