+++ app/index.tsx
@@ -1,12 +1,8 @@
 import { Text, View } from 'react-native';
-import { useState, useEffect } from 'react';
+import { useEffect } from 'react';
 import { router } from 'expo-router';
-import { TokenContext } from '../contexts/TokenContext';
 
 export default function App() {
-  const [accessToken, setAccessToken] = useState('');
-  const [refreshToken, setRefreshToken] = useState('');
-
   useEffect(() => {
     const timer = setTimeout(() => {
       router.push('/(auth)/sign-in');
@@ -15,10 +11,8 @@ export default function App() {
     return () => clearTimeout(timer);
   }, []);
   return (
-    <TokenContext.Provider value={{ accessToken, setAccessToken, refreshToken, setRefreshToken }}>
-      <View className="w-screen h-screen items-center justify-center bg-white font-sans-md">
-        <Text>GREET!</Text>
-      </View>
-    </TokenContext.Provider>
+    <View className="w-screen h-screen items-center justify-center bg-white font-sans-md">
+      <Text>GREET!</Text>
+    </View>
   );
 }
